requirements.txt
Flask>=2.0
azure-storage-blob>=12.8.0
python-dotenv>=0.19.0
